Define dev firewall before main firewall
This disables security for /_profiler and /_wdt. This means these URLs are now accessible when logged out, which also means the Symfony Web Toolbar will be shown when logged out.

"The order is significant because when making a request, the first matched firewall will be the one used.

If the patterns for the first firewall is ^/, then it will match all requests, and no other firewall will be evaluated."

https://stackoverflow.com/questions/58584088/is-the-firewall-declaration-order-significant-in-security-yaml